third_party/WebKit/LayoutTests/imported/wpt/service-workers/service-worker/fetch-cors-xhr.https-expected.txt

CONSOLE ERROR: line 177: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('https://www1.127.0.0.1:8444') does not match the recipient window's origin ('null').
This is a testharness.js-based test.
Harness Error. harness_status.status = 2 , harness_status.message = null
TIMEOUT Verify CORS XHR of fetch() in a Service Worker Test timed out
Harness: the test ran to completion.
